requirements.txt
python-dotenv
openai==1.35.14
openai-whisper
pyaudio
wave
numpy
sounddevice
soundfile
faiss-cpu